Added documentation for new map generation
